Abstract ServoQueueItem into own class
ConnectServo and ServoQueueItem reference each other, so the latter needs to forward reference the former. Tidies things up, too.
