+++ ConnectServo.cpp
@@ -1,18 +1,5 @@
 #include <ConnectServo.h>
 
-// ServoQueueItem initializer
-ServoQueueItem::ServoQueueItem() {
-
-};
-
-void ServoQueueItem::assign(uint8_t newCall, uint8_t newParam1, uint8_t newAnimationType, uint16_t newServoSpeed, ConnectServo* targetServo) {
-    call = newCall;
-    param1 = newParam1;
-    animationType = newAnimationType;
-    servoSpeed = newServoSpeed;
-    targetServo = targetServo;
-}
-
 // Initialize the servo object, passing an initializer list to the cppQueue object
 // See http://arduinoetcetera.blogspot.com/2011/01/classes-within-classes-initialiser.html
 ConnectServo::ConnectServo() : _servoQueue(sizeof(ServoQueueItem), QUEUE_SIZE_ITEMS, IMPLEMENTATION) {
