+++ ConnectServo.h
@@ -1,6 +1,7 @@
 #include <Arduino.h>
 #include <ServoEasing.h>
 #include <cppQueue.h>
+#include <ServoQueueItem.h>
 
 #define IMPLEMENTATION FIFO
 #define QUEUE_SIZE_ITEMS 20
@@ -11,31 +12,6 @@
 #define WAIT_FOR_OTHER_SERVO 2
 #define WAIT_FOR_LEDS 3
 #define MESSAGE_SERVO 4
-
-// Generic function pointer. Note variable-length argument list
-// typedef void (*GenericFP)(void *, ...);
-
-// Servo queue item type
-// typedef struct {
-//     GenericFP call;             // eg. startEaseTo, moveTo/write, waitForServo
-//     int param1;                 // Target position or (for waitForServo) servo to wait on.
-//     const char * animationType; // eg. 'EASE_CUBIC_IN_OUT'
-//     int servoSpeed;             // Slew rate (perscentage: map to 0-255. Servo theoretical max is around 200 degrees/sec, ~80%)
-// } servoQueueItem;
-
-
-
-class ServoQueueItem {
-    public:
-        ServoQueueItem();
-        // call, target position, animation type, servo speed
-        void assign(uint8_t, uint8_t, uint8_t, uint16_t, ConnectServo*);
-        uint8_t call;                   // int representing call; we'll parse in ConnectServo::update()
-        uint8_t param1;                 // Target position or (for waitForServo) servo to wait on.
-        uint8_t animationType; // eg. 'EASE_CUBIC_IN_OUT'
-        uint16_t servoSpeed;             // Slew rate (perscentage: map to 0-255. Servo theoretical max is around 200 degrees/sec, ~80%)
-        ConnectServo* targetServo;
-};
 class ConnectServo : public ServoEasing {
     public:
         ConnectServo();
